+++ lib/src/core/models/remove_customer/remove_customer_response.dart
@@ -33,9 +33,11 @@ class RemoveCustomerResponse extends AcquiringResponse {
   Map<String, dynamic> toJson() => _$RemoveCustomerResponseToJson(this);
 
   @override
-  String toString() {
-    return 'RemoveCustomerResponse(terminalKey: $terminalKey, customerKey: $customerKey, success: $success, errorCode: $errorCode, message: $message, details: $details)';
-  }
+  Map<String, Object> get equals => <String, Object>{
+        ...super.equals,
+        JsonKeys.terminalKey: terminalKey,
+        JsonKeys.customerKey: customerKey,
+      };
 
   /// Идентификатор терминала.
   /// Выдается продавцу банком при заведении терминала