+++ lib/src/core/utils/extensions.dart
@@ -1,5 +1,11 @@
-/// Get length method in num type
-extension Num on num {
+/// Extensions for [num]
+extension NumEx on num {
   /// length in num type
   int get length => toString().length;
+}
+
+/// Extensions for [Map]
+extension MapEx<K, V> on Map<K, V> {
+  /// normalize string
+  String toNormalizeString() => toString().replaceAll(RegExp(r'^{|}$'), '');
 }