+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+* Add Comparer (`hashCode`, `==` & `toString` generator)
+* Fixed example using 3-D Secure 2.0
+
 # 1.0.1
 
 * Add support 3-D Secure 2.0